Nvidia version fixes

+++ protocol/fractal/video/nvidia-linux/NvFBCUtils.c
@@ -35,59 +35,49 @@
 #define BITMAP_ROW_SIZE(width) (((width * Bpp) + 3) & ~3)
 #define BITMAP_INDEX(x, y, rowSize) ((y * rowSize) + (x * Bpp))
 
-typedef struct __attribute__((packed))
-{
+typedef struct __attribute__((packed)) {
     uint8_t blue;
     uint8_t green;
     uint8_t red;
 } BitmapPixel;
 
-typedef struct __attribute__((packed))
-{
+typedef struct __attribute__((packed)) {
     uint8_t alpha;
     uint8_t red;
     uint8_t green;
     uint8_t blue;
 } ARGBPixel;
 
-typedef struct __attribute__((packed))
-{
+typedef struct __attribute__((packed)) {
     uint8_t red;
     uint8_t green;
     uint8_t blue;
     uint8_t alpha;
 } RGBAPixel;
 
-typedef struct __attribute__((packed))
-{
+typedef struct __attribute__((packed)) {
     uint8_t blue;
     uint8_t green;
     uint8_t red;
     uint8_t alpha;
 } BGRAPixel;
 
-typedef struct __attribute__((packed))
-{
+typedef struct __attribute__((packed)) {
     uint8_t red;
     uint8_t green;
     uint8_t blue;
 } RGBPixel;
 
-typedef struct __attribute__((packed))
-{
+typedef struct __attribute__((packed)) {
     uint8_t red;
 } RPixel;
 
 // Bytes per pixel
 static int Bpp = sizeof(BitmapPixel);
 
-static int NvFBCUtilsSaveBitmap(const char *filename,
-                                const unsigned char *data,
-                                const int width,
-                                const int height)
-{
-    struct __attribute__((packed))
-    {
+static int NvFBCUtilsSaveBitmap(const char *filename, const unsigned char *data, const int width,
+                                const int height) {
+    struct __attribute__((packed)) {
         uint16_t type;
         uint32_t size;
         uint16_t reserved1;
@@ -95,17 +85,16 @@ static int NvFBCUtilsSaveBitmap(const char *filename,
         uint32_t off_bits;
     } fileHeader;
 
-    struct __attribute__((packed))
-    {
+    struct __attribute__((packed)) {
         uint32_t size;
-        int32_t  width;
-        int32_t  height;
+        int32_t width;
+        int32_t height;
         uint16_t planes;
         uint16_t bit_count;
         uint32_t compression;
         uint32_t size_image;
-        int32_t  x_pels_per_meter;
-        int32_t  y_pels_per_meter;
+        int32_t x_pels_per_meter;
+        int32_t y_pels_per_meter;
         uint32_t clr_used;
         uint32_t clr_important;
     } infoHeader;
@@ -126,24 +115,24 @@ static int NvFBCUtilsSaveBitmap(const char *filename,
 
     size = BITMAP_ROW_SIZE(width) * height;
 
-    fileHeader.type     = 0x4D42;
-    fileHeader.size     = sizeof(fileHeader) + sizeof(infoHeader) + size;
+    fileHeader.type = 0x4D42;
+    fileHeader.size = sizeof(fileHeader) + sizeof(infoHeader) + size;
     fileHeader.off_bits = sizeof(fileHeader) + sizeof(infoHeader);
 
-    infoHeader.size             = sizeof(infoHeader);
-    infoHeader.width            = width;
-    infoHeader.height           = height;
-    infoHeader.planes           = 1;
-    infoHeader.bit_count        = Bpp * 8;
-    infoHeader.compression      = 0;
-    infoHeader.size_image       = 0;
+    infoHeader.size = sizeof(infoHeader);
+    infoHeader.width = width;
+    infoHeader.height = height;
+    infoHeader.planes = 1;
+    infoHeader.bit_count = Bpp * 8;
+    infoHeader.compression = 0;
+    infoHeader.size_image = 0;
     infoHeader.x_pels_per_meter = 0;
     infoHeader.y_pels_per_meter = 0;
-    infoHeader.clr_used         = 0;
-    infoHeader.clr_important    = 0;
+    infoHeader.clr_used = 0;
+    infoHeader.clr_important = 0;
 
-    fwrite((unsigned char *) &fileHeader, sizeof(fileHeader), 1, fd);
-    fwrite((unsigned char *) &infoHeader, sizeof(infoHeader), 1, fd);
+    fwrite((unsigned char *)&fileHeader, sizeof(fileHeader), 1, fd);
+    fwrite((unsigned char *)&infoHeader, sizeof(infoHeader), 1, fd);
     fwrite(data, size, 1, fd);
 
     fclose(fd);
@@ -151,12 +140,8 @@ static int NvFBCUtilsSaveBitmap(const char *filename,
     return 0;
 }
 
-static int NvFBCUtilsSaveRGBFormats(NVFBC_BUFFER_FORMAT format,
-                                    const char *filename,
-                                    unsigned char *data,
-                                    const int width,
-                                    const int height)
-{
+static int NvFBCUtilsSaveRGBFormats(NVFBC_BUFFER_FORMAT format, const char *filename,
+                                    unsigned char *data, const int width, const int height) {
     int result = 1;
 
     int row, col;
@@ -208,9 +193,9 @@ static int NvFBCUtilsSaveRGBFormats(NVFBC_BUFFER_FORMAT format,
             }
 
             BitmapPixel *pixelOut = (BitmapPixel *)&output[outputIdx];
-            pixelOut->red   = r;
+            pixelOut->red = r;
             pixelOut->green = g;
-            pixelOut->blue  = b;
+            pixelOut->blue = b;
         }
     }
 
@@ -221,11 +206,8 @@ static int NvFBCUtilsSaveRGBFormats(NVFBC_BUFFER_FORMAT format,
     return result;
 }
 
-int NvFBCUtilsSaveDiffMap(const char *filename,
-                          unsigned char *data,
-                          const int width,
-                          const int height)
-{
+int NvFBCUtilsSaveDiffMap(const char *filename, unsigned char *data, const int width,
+                          const int height) {
     int result = 1;
 
     int row, col;
@@ -240,18 +222,16 @@ int NvFBCUtilsSaveDiffMap(const char *filename,
     // Pad bytes need to be set to zero, it's easier to just set the entire chunk of memory
     memset(output, 0, size);
 
-    for (row = 0; row < height; ++row)
-    {
-        for (col = 0; col < width; ++col)
-        {
+    for (row = 0; row < height; ++row) {
+        for (col = 0; col < width; ++col) {
             // In a bitmap (0,0) is at the bottom left, in the frame buffer it is the top left.
             int outputIdx = BITMAP_INDEX(col, row, rowSize);
             int inputIdx = ((height - row - 1) * width) + col;
 
             BitmapPixel *pixelOut = (BitmapPixel *)&output[outputIdx];
-            pixelOut->red   = input[inputIdx].red;
+            pixelOut->red = input[inputIdx].red;
             pixelOut->green = input[inputIdx].red;
-            pixelOut->blue  = input[inputIdx].red;
+            pixelOut->blue = input[inputIdx].red;
         }
     }
 
@@ -262,12 +242,8 @@ int NvFBCUtilsSaveDiffMap(const char *filename,
     return result;
 }
 
-static int NvFBCUtilsSaveYUVPlanar(NVFBC_BUFFER_FORMAT format,
-                                   const char *fileName,
-                                   unsigned char *data,
-                                   const int width,
-                                   const int height)
-{
+static int NvFBCUtilsSaveYUVPlanar(NVFBC_BUFFER_FORMAT format, const char *fileName,
+                                   unsigned char *data, const int width, const int height) {
     int result = 1;
 
     int row, col;
@@ -369,12 +345,8 @@ done:
     return result;
 }
 
-int NvFBCUtilsSaveFrame(NVFBC_BUFFER_FORMAT format,
-                        const char *filename,
-                        unsigned char *data,
-                        const int width,
-                        const int height)
-{
+int NvFBCUtilsSaveFrame(NVFBC_BUFFER_FORMAT format, const char *filename, unsigned char *data,
+                        const int width, const int height) {
     switch (format) {
         case NVFBC_BUFFER_FORMAT_ARGB:
         case NVFBC_BUFFER_FORMAT_RGBA:
@@ -390,8 +362,7 @@ int NvFBCUtilsSaveFrame(NVFBC_BUFFER_FORMAT format,
     }
 }
 
-uint64_t NvFBCUtilsGetTimeInMicros()
-{
+uint64_t NvFBCUtilsGetTimeInMicros() {
     struct timeval tv;
 
     gettimeofday(&tv, NULL);
@@ -399,55 +370,41 @@ uint64_t NvFBCUtilsGetTimeInMicros()
     return (uint64_t)tv.tv_sec * 1000000ULL + (uint64_t)tv.tv_usec;
 }
 
-uint64_t NvFBCUtilsGetTimeInMillis()
-{
-    return NvFBCUtilsGetTimeInMicros() / 1000;
-}
+uint64_t NvFBCUtilsGetTimeInMillis() { return NvFBCUtilsGetTimeInMicros() / 1000; }
 
-void NvFBCUtilsPrintStatus(NVFBC_GET_STATUS_PARAMS *status)
-{
+void NvFBCUtilsPrintStatus(NVFBC_GET_STATUS_PARAMS *status) {
     if (status == NULL) {
         return;
     }
 
     printf("Status:\n");
-    printf("- NvFBC library API version: %u.%u\n",
-           status->dwNvFBCVersion >> 8 & 0xf,
+    printf("- NvFBC library API version: %u.%u\n", status->dwNvFBCVersion >> 8 & 0xf,
            status->dwNvFBCVersion & 0xf);
-    printf("- This system supports FBC: %s\n",
-           status->bIsCapturePossible ? "Yes" : "No");
-    printf("- Curently capturing: %s\n",
-           status->bCurrentlyCapturing ? "Yes" : "No");
-    printf("- Can create an FBC instance: %s\n",
-           status->bCanCreateNow ? "Yes" : "No");
-    printf("- X screen (framebuffer) size: %ux%u\n",
-           status->screenSize.w, status->screenSize.h);
-    printf("- XrandR extension available: %s\n",
-           status->bXRandRAvailable ? "Yes" : "No");
+    printf("- This system supports FBC: %s\n", status->bIsCapturePossible ? "Yes" : "No");
+    printf("- Curently capturing: %s\n", status->bCurrentlyCapturing ? "Yes" : "No");
+    printf("- Can create an FBC instance: %s\n", status->bCanCreateNow ? "Yes" : "No");
+    printf("- X screen (framebuffer) size: %ux%u\n", status->screenSize.w, status->screenSize.h);
+    printf("- XrandR extension available: %s\n", status->bXRandRAvailable ? "Yes" : "No");
 
     if (status->bXRandRAvailable) {
         int i;
 
         printf("- Connected RandR outputs with CRTC:\n");
 
-        for (i = 0; i < status->dwOutputNum; i++) {
+        for (i = 0; i < (int)status->dwOutputNum; i++) {
             NVFBC_RANDR_OUTPUT_INFO output;
 
             output = status->outputs[i];
 
-            printf("  * '%s' (id: 0x%x), CRTC: %ux%u+%u+%u\n",
-                   output.name, output.dwId,
-                   output.trackedBox.w, output.trackedBox.h,
-                   output.trackedBox.x, output.trackedBox.y);
+            printf("  * '%s' (id: 0x%x), CRTC: %ux%u+%u+%u\n", output.name, output.dwId,
+                   output.trackedBox.w, output.trackedBox.h, output.trackedBox.x,
+                   output.trackedBox.y);
         }
-
     }
 }
 
-void NvFBCUtilsParseTrackingType(const char* optarg,
-                                 NVFBC_TRACKING_TYPE *trackingType,
-                                 char* outputName)
-{
+void NvFBCUtilsParseTrackingType(const char *optarg, NVFBC_TRACKING_TYPE *trackingType,
+                                 char *outputName) {
     if ((trackingType == NULL) || (outputName == NULL)) {
         return;
     }
@@ -462,10 +419,8 @@ void NvFBCUtilsParseTrackingType(const char* optarg,
     }
 }
 
-uint32_t NvFBCUtilsGetOutputId(NVFBC_RANDR_OUTPUT_INFO *outputs,
-                               uint32_t outputNum,
-                               const char* outputName)
-{
+uint32_t NvFBCUtilsGetOutputId(NVFBC_RANDR_OUTPUT_INFO *outputs, uint32_t outputNum,
+                               const char *outputName) {
     int i;
     uint32_t outputId = 0;
 
@@ -473,7 +428,7 @@ uint32_t NvFBCUtilsGetOutputId(NVFBC_RANDR_OUTPUT_INFO *outputs,
         return 0;
     }
 
-    for (i = 0; i < outputNum; i++) {
+    for (i = 0; i < (int)outputNum; i++) {
         if (!strcasecmp(outputs[i].name, outputName)) {
             outputId = outputs[i].dwId;
             break;
@@ -483,8 +438,7 @@ uint32_t NvFBCUtilsGetOutputId(NVFBC_RANDR_OUTPUT_INFO *outputs,
     return outputId;
 }
 
-void NvFBCUtilsPrintVersions(const unsigned int appVersion)
-{
+void NvFBCUtilsPrintVersions(const unsigned int appVersion) {
     printf("Application version: %u\n", appVersion);
     printf("NvFBC API version: %u.%u\n", NVFBC_VERSION_MAJOR, NVFBC_VERSION_MINOR);
     printf("\n");
